refactor(parser): split tokenization and structured reuse
