+++ packages/markdown-parser/src/parser/index.ts
@@ -3,11 +3,6 @@ import type { HtmlBlockNode, InternalParseOptions, MarkdownToken, ParsedNode, Pa
 import { normalizeCustomHtmlTags } from '../customHtmlTags'
 import { NON_STRUCTURING_HTML_TAGS, VOID_HTML_TAGS } from '../htmlTags'
 import { escapeTagForRegExp, findTagCloseIndexOutsideQuotes, parseTagAttrs } from '../htmlTagUtils'
-import { isCacheStableLinkValidator, readSyntheticLinkOrigin } from '../plugins/linkTokenMetadata'
-import {
-  hasMarkstreamMathPlugin,
-  mayContainTolerantMathBlockBoundaryOpener,
-} from '../plugins/math'
 import { parseInlineTokens } from './inline-parsers'
 import { createLinkifyDemotionContextTracker } from './linkifyHeuristics'
 import { parseCommonBlockToken } from './node-parsers/block-token-parser'
@@ -18,14 +13,14 @@ import { parseHtmlBlock } from './node-parsers/html-block-parser'
 import { parseList } from './node-parsers/list-parser'
 import { parseParagraph } from './node-parsers/paragraph-parser'
 import { applyNodeSourceMap, createSourceMapFromOffsets } from './node-source-map'
+import { processTopLevelTokensWithReuse } from './reuse/structured-node-reuse'
 import { createSourceLineMapper } from './source-line-mapper'
-import {
-  clearTolerantMathBoundaryStreamCache,
-  shouldUseSyncParseForPendingTolerantMathBoundary,
-  syncTolerantMathBoundaryStreamCache,
-} from './streaming/boundary-state'
 import { clearSafeMarkdownCache, getSafeMarkdown } from './streaming/safe-markdown'
-import { cloneMarkdownTokens } from './token-clone'
+import {
+  parseTopLevelTokens,
+  resetTopLevelTokenizerForFinalAutoParse,
+  shouldUseTopLevelStreamParse,
+} from './streaming/tokenizer'
 
 type ParsedNodeWithFields = ParsedNode & {
   children?: ParsedNode[]
@@ -33,7 +28,6 @@ type ParsedNodeWithFields = ParsedNode & {
   tag?: unknown
 }
 
-const streamParseEnvCache = new WeakMap<object, Map<string, Record<string, unknown>>>()
 const internalNodeSourceRanges = new WeakMap<object, { start: number, end: number }>()
 const sourceLineOffsetsCache = new WeakMap<object, number[]>()
 const siblingHtmlChildrenCache = new WeakMap<object, {
@@ -86,70 +80,6 @@ interface ParseTimingMetrics {
   htmlBlockPassesMs?: number
   parseMarkdownToStructureTotalMs?: number
 }
-
-interface StructuredStreamCacheEntry {
-  groupBoundaries: StructuredStreamGroupBoundary[]
-  source: string
-  nodes: ParsedNode[]
-  stableGroupCount: number
-  requireClosingStrong: boolean | undefined
-  validateLink: ParseOptions['validateLink']
-}
-
-interface StructuredStreamGroupBoundary {
-  firstToken: MarkdownToken
-  lastToken: MarkdownToken
-  tokenCount: number
-}
-
-interface ReusableTopLevelTokenGroups {
-  mixed: boolean
-  starts: number[]
-}
-
-const structuredStreamCache = new WeakMap<object, StructuredStreamCacheEntry>()
-const topLevelStreamParseMode = new WeakMap<object, string>()
-const REUSABLE_INLINE_TOKEN_TYPES = new Set([
-  'code_inline',
-  'em_close',
-  'em_open',
-  'emoji',
-  'hardbreak',
-  'html_block',
-  'html_inline',
-  'image',
-  'ins_close',
-  'ins_open',
-  'link',
-  'link_close',
-  'link_open',
-  'mark_close',
-  'mark_open',
-  'math_inline',
-  's_close',
-  's_open',
-  'softbreak',
-  'strong_close',
-  'strong_open',
-  'sub',
-  'sup',
-  'text',
-])
-const REUSABLE_TOP_LEVEL_PAIRED_TOKEN_TYPES = new Map([
-  ['paragraph_open', 'paragraph_close'],
-  ['heading_open', 'heading_close'],
-  ['bullet_list_open', 'bullet_list_close'],
-  ['ordered_list_open', 'ordered_list_close'],
-  ['blockquote_open', 'blockquote_close'],
-  ['table_open', 'table_close'],
-])
-const REUSABLE_TOP_LEVEL_SINGLE_TOKEN_TYPES = new Set([
-  'code_block',
-  'fence',
-  'hr',
-  'inline',
-  'math_block',
-])
 
 type TimedParseOptions = ParseOptions & {
   __timing?: ParseTimingMetrics
@@ -212,283 +142,6 @@ function processTokensWithTiming(tokens: MarkdownToken[], options: ParseOptions 
   return result
 }
 
-function hasOnlyReusableInlineTokens(tokens: MarkdownToken[], validateLink: ParseOptions['validateLink']): boolean {
-  return tokens.every((token) => {
-    if (!REUSABLE_INLINE_TOKEN_TYPES.has(token.type))
-      return false
-    if (
-      (token.type === 'link' || token.type === 'link_open' || token.type === 'link_close')
-      && !isCacheStableLinkValidator(validateLink)
-    ) {
-      return false
-    }
-    if (token.type === 'link') {
-      // fixLinkTokens emits `link` single tokens with a recorded origin.
-      // explicit/linkify/autolink origins produce nodes deterministically
-      // from the token + its inline group (recovery paths are group-local);
-      // `recovery` tokens are emitted for broken streaming link tails and
-      // stay excluded.
-      const origin = readSyntheticLinkOrigin(token)
-      if (origin !== 'explicit' && origin !== 'linkify' && origin !== 'autolink')
-        return false
-    }
-    if (token.type === 'link_open' || token.type === 'link_close') {
-      // Explicit links carry an empty markup. markdown-it linkify emits
-      // `link_open`/`link_close` pairs with markup `linkify`/`autolink`;
-      // their node output is deterministic given the token, and the tail
-      // re-parse seeds the linkify demotion tracker with the reused prefix
-      // context, so these pairs are safe to reuse.
-      const markup = token.markup ?? ''
-      if (markup !== '' && markup !== 'linkify' && markup !== 'autolink')
-        return false
-    }
-
-    const children = token.children as MarkdownToken[] | null
-    return !Array.isArray(children) || hasOnlyReusableInlineTokens(children, validateLink)
-  })
-}
-
-function getReusableTopLevelPairedCloseType(tokenType: string) {
-  const known = REUSABLE_TOP_LEVEL_PAIRED_TOKEN_TYPES.get(tokenType)
-  if (known)
-    return known
-
-  // markdown-it-container emits `container_<kind>_open` / `container_<kind>_close`
-  // at level 0 for every registered container name. The node output is a pure
-  // function of the token pair (attrs/info + children), so these groups are as
-  // reusable as the statically known pairs.
-  const containerMatch = /^container_(.+)_open$/.exec(tokenType)
-  return containerMatch ? `container_${containerMatch[1]}_close` : undefined
-}
-
-function getReusableTopLevelTokenGroups(
-  tokens: MarkdownToken[],
-  validateLink: ParseOptions['validateLink'],
-): ReusableTopLevelTokenGroups | null {
-  const groupStarts: number[] = []
-  let mixed = false
-  let index = 0
-
-  while (index < tokens.length) {
-    const token = tokens[index]
-    if (!token || token.level !== 0)
-      return null
-
-    const closeType = getReusableTopLevelPairedCloseType(token.type)
-    let groupEnd = index + 1
-
-    if (closeType) {
-      if (token.nesting !== 1)
-        return null
-
-      while (groupEnd < tokens.length) {
-        const current = tokens[groupEnd]
-        if (current.level === 0) {
-          if (current.type !== closeType || current.nesting !== -1)
-            return null
-          groupEnd++
-          break
-        }
-        groupEnd++
-      }
-
-      if (tokens[groupEnd - 1]?.type !== closeType)
-        return null
-
-      if (token.type === 'paragraph_open' || token.type === 'heading_open') {
-        if (groupEnd !== index + 3 || tokens[index + 1]?.type !== 'inline')
-          return null
-      }
-      else {
-        mixed = true
-      }
-    }
-    else if (REUSABLE_TOP_LEVEL_SINGLE_TOKEN_TYPES.has(token.type)) {
-      if (token.nesting !== 0)
-        return null
-      mixed = true
-    }
-    else {
-      return null
-    }
-
-    for (let tokenIndex = index; tokenIndex < groupEnd; tokenIndex++) {
-      const current = tokens[tokenIndex]
-      if (current.type !== 'inline')
-        continue
-      const children = current.children as MarkdownToken[] | null
-      if (!Array.isArray(children) || !hasOnlyReusableInlineTokens(children, validateLink))
-        return null
-    }
-
-    groupStarts.push(index)
-    index = groupEnd
-  }
-
-  return {
-    mixed,
-    starts: groupStarts,
-  }
-}
-
-function sourceEndsWithBlankLine(source: string) {
-  return /\r?\n[\t ]*\r?\n[\t ]*$/.test(source)
-}
-
-function canReuseStructuredStreamNodes(options: ParseOptions) {
-  return (options as InternalParseOptions).__reuseStableTopLevelNodes === true
-    && options.final !== true
-    && !options.preTransformTokens
-    && !options.postTransformTokens
-    && !options.postTransformNodes
-    && !options.customHtmlTags?.length
-    && options.includeSourceMap !== true
-}
-
-function updateStructuredStreamCache(
-  md: MarkdownIt,
-  source: string,
-  tokens: MarkdownToken[],
-  groups: ReusableTopLevelTokenGroups,
-  nodes: ParsedNode[],
-  options: ParseOptions,
-) {
-  const groupStarts = groups.starts
-  if (groupStarts.length === 0 || nodes.length !== groupStarts.length) {
-    structuredStreamCache.delete(md as unknown as object)
-    return
-  }
-
-  const groupBoundaries = groupStarts.map((start, index) => {
-    const end = groupStarts[index + 1] ?? tokens.length
-    return {
-      firstToken: tokens[start],
-      lastToken: tokens[end - 1],
-      tokenCount: end - start,
-    }
-  })
-
-  structuredStreamCache.set(md as unknown as object, {
-    groupBoundaries,
-    source,
-    nodes,
-    stableGroupCount: groups.mixed
-      ? Math.max(0, groupStarts.length - 1)
-      : sourceEndsWithBlankLine(source) ? groupStarts.length : Math.max(0, groupStarts.length - 1),
-    requireClosingStrong: options.requireClosingStrong,
-    validateLink: options.validateLink,
-  })
-}
-
-function hasStableStructuredStreamGroupBoundaries(
-  previous: StructuredStreamCacheEntry,
-  tokens: MarkdownToken[],
-  groupStarts: number[],
-  stableGroupCount: number,
-) {
-  const lastGroupIndex = groupStarts.length - 1
-  for (let index = 0; index < stableGroupCount; index++) {
-    const start = groupStarts[index]
-    const end = groupStarts[index + 1] ?? tokens.length
-    const boundary = previous.groupBoundaries[index]
-    if (
-      !boundary
-      || boundary.tokenCount !== end - start
-    ) {
-      return false
-    }
-    const identical = boundary.firstToken === tokens[start]
-      && boundary.lastToken === tokens[end - 1]
-    if (identical)
-      continue
-    // The stream parser can recreate prefix tokens on a full re-parse or a
-    // container tail merge. A deterministic re-parse of unchanged source
-    // produces shape-equal tokens, so boundary shape equality proves the
-    // group content is unchanged — EXCEPT for the last group, which may
-    // legitimately gain new content while its boundary tokens keep the same
-    // shape (e.g. a tail merge that recreates the open/close pair). Interior
-    // groups can never receive appended content, so shape fallback is safe
-    // only below the last group.
-    if (index >= lastGroupIndex || !isSameTokenShapeForReuse(boundary.firstToken, tokens[start])
-      || !isSameTokenShapeForReuse(boundary.lastToken, tokens[end - 1])) {
-      return false
-    }
-  }
-
-  return true
-}
-
-function processTopLevelTokensWithReuse(
-  md: MarkdownIt,
-  source: string,
-  tokens: MarkdownToken[],
-  options: InternalParseOptions,
-  timing: ParseTimingMetrics | undefined,
-) {
-  const owner = md as unknown as object
-  const structuredReuseDisabled = (options as InternalParseOptions).__disableStructuredReuse === true
-  const reuseEnabled = shouldUseTopLevelStreamParse(md, options)
-    && canReuseStructuredStreamNodes(options)
-
-  if (!reuseEnabled) {
-    // Fragment parses (e.g. children of <details>/custom html blocks) run with
-    // the same md instance and must not evict the top-level document's
-    // structured reuse cache.
-    if (!structuredReuseDisabled)
-      structuredStreamCache.delete(owner)
-    return processTokensWithTiming(tokens, options, timing)
-  }
-
-  if (structuredReuseDisabled)
-    return processTokensWithTiming(tokens, options, timing)
-
-  const groups = getReusableTopLevelTokenGroups(tokens, options.validateLink)
-  if (!groups) {
-    structuredStreamCache.delete(owner)
-    return processTokensWithTiming(tokens, options, timing)
-  }
-
-  const groupStarts = groups.starts
-  const previous = structuredStreamCache.get(owner)
-  const mode = topLevelStreamParseMode.get(owner)
-  const stableGroupCount = previous && groups.mixed
-    ? Math.min(previous.stableGroupCount, Math.max(0, previous.groupBoundaries.length - 1))
-    : previous?.stableGroupCount ?? 0
-  if (
-    previous
-    && stableGroupCount > 0
-    && previous.requireClosingStrong === options.requireClosingStrong
-    && previous.validateLink === options.validateLink
-    && source.startsWith(previous.source)
-    && groupStarts.length >= stableGroupCount
-    && (mode === 'append' || mode === 'tail')
-    && hasStableStructuredStreamGroupBoundaries(previous, tokens, groupStarts, stableGroupCount)
-  ) {
-    const tailStart = groupStarts[stableGroupCount] ?? tokens.length
-    const tailNodes = processTokensWithTiming(tokens.slice(tailStart), {
-      ...options,
-      // Replay the reused prefix node raws into the tail's linkify demotion
-      // tracker so tail linkify decisions see the same accumulated context a
-      // full parse would have produced.
-      __linkifyDemotionSeed: previous.nodes
-        .slice(0, stableGroupCount)
-        .map(node => String((node as Record<string, unknown>).raw ?? '')),
-    } as InternalParseOptions, timing)
-    const expectedTailNodes = groupStarts.length - stableGroupCount
-
-    if (tailNodes.length === expectedTailNodes) {
-      const result = previous.nodes.slice(0, stableGroupCount).concat(tailNodes)
-      addTiming(timing, 'processTokensReusedTopLevelNodes', stableGroupCount)
-      updateStructuredStreamCache(md, source, tokens, groups, result, options)
-      return result
-    }
-  }
-
-  const result = processTokensWithTiming(tokens, options, timing)
-  updateStructuredStreamCache(md, source, tokens, groups, result, options)
-  return result
-}
-
 function getCustomHtmlTagSet(options?: ParseOptions) {
   const custom = options?.customHtmlTags
   if (!Array.isArray(custom) || custom.length === 0)
@@ -496,211 +149,6 @@ function getCustomHtmlTagSet(options?: ParseOptions) {
 
   const normalized = normalizeCustomHtmlTags(custom)
   return normalized.length ? new Set(normalized) : null
-}
-
-function getStableStreamEnv(md: MarkdownIt, env: Record<string, unknown>) {
-  const mdKey = md as unknown as object
-  let byMode = streamParseEnvCache.get(mdKey)
-  if (!byMode) {
-    byMode = new Map()
-    streamParseEnvCache.set(mdKey, byMode)
-  }
-
-  const modeKey = env.__markstreamFinal === true ? 'final' : 'streaming'
-  let stableEnv = byMode.get(modeKey)
-  if (!stableEnv) {
-    stableEnv = {}
-    byMode.set(modeKey, stableEnv)
-  }
-
-  for (const key of Object.keys(stableEnv)) {
-    if (!Object.prototype.hasOwnProperty.call(env, key))
-      delete stableEnv[key]
-  }
-  Object.assign(stableEnv, env)
-  return stableEnv
-}
-
-function shouldUseTopLevelStreamParse(md: MarkdownIt, options: ParseOptions) {
-  const internalOptions = options as InternalParseOptions
-  const stream = md.stream
-  const streamParse = options.streamParse ?? 'auto'
-  return internalOptions.__disableStreamParse !== true
-    && (md as unknown as Record<string, unknown>).__markstreamHasCustomParserExtensions !== true
-    && (streamParse === true || (streamParse === 'auto' && options.final !== true))
-    && stream?.enabled === true
-    && typeof stream.parse === 'function'
-}
-
-function shouldResetTopLevelStreamCacheForFinalAutoParse(md: MarkdownIt, options: ParseOptions) {
-  const internalOptions = options as InternalParseOptions
-  const streamParse = options.streamParse ?? 'auto'
-  const stream = md.stream
-
-  return options.final === true
-    && streamParse === 'auto'
-    && internalOptions.__disableStreamParse !== true
-    && (md as unknown as Record<string, unknown>).__markstreamHasCustomParserExtensions !== true
-    && stream?.enabled === true
-    && typeof stream.reset === 'function'
-}
-
-function shouldCloneTopLevelStreamTokens(options: ParseOptions) {
-  return typeof options.preTransformTokens === 'function'
-    || typeof options.postTransformTokens === 'function'
-}
-
-function sameTokenMap(left: Token | MarkdownToken | undefined, right: Token | MarkdownToken | undefined) {
-  const leftMap = left?.map
-  const rightMap = right?.map
-
-  if (leftMap === rightMap)
-    return true
-
-  if (!Array.isArray(leftMap) || !Array.isArray(rightMap))
-    return false
-
-  return leftMap.length === rightMap.length
-    && leftMap.every((value, index) => value === rightMap[index])
-}
-
-function sameTokenAttrs(left: Token | MarkdownToken | undefined, right: Token | MarkdownToken | undefined) {
-  const leftAttrs = left?.attrs
-  const rightAttrs = right?.attrs
-
-  if (leftAttrs === rightAttrs)
-    return true
-
-  if (!Array.isArray(leftAttrs) || !Array.isArray(rightAttrs))
-    return false
-
-  if (leftAttrs.length !== rightAttrs.length)
-    return false
-
-  for (let index = 0; index < leftAttrs.length; index++) {
-    const leftAttr = leftAttrs[index]
-    const rightAttr = rightAttrs[index]
-    if (leftAttr[0] !== rightAttr[0] || leftAttr[1] !== rightAttr[1])
-      return false
-  }
-
-  return true
-}
-
-/**
- * Shape equality for reuse-boundary tokens, used when the stream parser
- * recreates prefix tokens after a full re-parse or a container tail merge
- * (identity comparison fails because the tokens are new objects).
- *
- * Correctness rests on markdown-it tokenization being a deterministic
- * function of the source text: identical source in the stream re-parse
- * yields tokens with identical shape INCLUDING fields not compared here
- * (children, meta, interior group tokens). The shape fallback therefore
- * never detects a change that identity comparison would have caught; it
- * only re-admits groups whose source provably did not change (interior
- * groups never receive appended content). `level` is compared because a
- * re-parsed boundary token at a different nesting depth cannot be the
- * same group.
- */
-function isSameTokenShapeForReuse(left: Token | MarkdownToken | undefined, right: Token | MarkdownToken | undefined) {
-  return !!left
-    && !!right
-    && left.type === right.type
-    && left.tag === right.tag
-    && left.nesting === right.nesting
-    && left.level === right.level
-    && left.markup === right.markup
-    && left.content === right.content
-    && left.info === right.info
-    && sameTokenMap(left, right)
-    && sameTokenAttrs(left, right)
-}
-
-function isSameTokenShape(left: Token | undefined, right: Token | undefined) {
-  return !!left
-    && !!right
-    && left.type === right.type
-    && left.tag === right.tag
-    && left.nesting === right.nesting
-    && left.markup === right.markup
-    && left.content === right.content
-    && sameTokenMap(left, right)
-}
-
-function isParagraphTokenTriplet(tokens: Token[], index: number) {
-  return tokens[index]?.type === 'paragraph_open'
-    && tokens[index + 1]?.type === 'inline'
-    && tokens[index + 2]?.type === 'paragraph_close'
-}
-
-function hasAdjacentDuplicateParagraphTokenTriplet(tokens: Token[]) {
-  for (let index = 0; index + 5 < tokens.length; index++) {
-    if (
-      isParagraphTokenTriplet(tokens, index)
-      && isParagraphTokenTriplet(tokens, index + 3)
-      && isSameTokenShape(tokens[index], tokens[index + 3])
-      && isSameTokenShape(tokens[index + 1], tokens[index + 4])
-      && isSameTokenShape(tokens[index + 2], tokens[index + 5])
-    ) {
-      return true
-    }
-  }
-
-  return false
-}
-
-function shouldFallbackDuplicateTolerantMathStreamTokens(
-  md: MarkdownIt,
-  source: string,
-  tokens: Token[],
-) {
-  return hasMarkstreamMathPlugin(md)
-    && mayContainTolerantMathBlockBoundaryOpener(source)
-    && hasAdjacentDuplicateParagraphTokenTriplet(tokens)
-}
-
-function parseTopLevelTokens(
-  md: MarkdownIt,
-  source: string,
-  env: Record<string, unknown>,
-  options: ParseOptions,
-) {
-  const owner = md as unknown as object
-  if (options.customHtmlTags?.length)
-    env.__markstreamCustomHtmlTags = options.customHtmlTags
-
-  if (!shouldUseTopLevelStreamParse(md, options)) {
-    topLevelStreamParseMode.set(owner, 'sync')
-    return md.parse(source, env)
-  }
-
-  syncTolerantMathBoundaryStreamCache(md, source)
-  if (shouldUseSyncParseForPendingTolerantMathBoundary(md)) {
-    topLevelStreamParseMode.set(owner, 'sync')
-    return md.parse(source, env)
-  }
-
-  const tokens = md.stream!.parse!(source, getStableStreamEnv(md, env))
-  if (shouldFallbackDuplicateTolerantMathStreamTokens(md, source, tokens)) {
-    md.stream?.reset?.()
-    topLevelStreamParseMode.set(owner, 'sync')
-    return md.parse(source, env)
-  }
-
-  const stats = md.stream?.stats?.() as { lastMode?: string } | undefined
-  topLevelStreamParseMode.set(owner, stats?.lastMode ?? 'stream')
-
-  if (!shouldCloneTopLevelStreamTokens(options))
-    return tokens
-
-  const timing = getParseTiming(options)
-  if (!timing)
-    return cloneMarkdownTokens(tokens, true)
-
-  const startedAt = getParserNow()
-  const cloned = cloneMarkdownTokens(tokens, true)
-  addTiming(timing, 'tokenCloneMs', getParserNow() - startedAt)
-  return cloned
 }
 
 function stringifyInlineNodeRaw(node: ParsedNode) {
@@ -1603,14 +1051,15 @@ export function parseMarkdownToStructure(
   options: ParseOptions = {},
 ): ParsedNode[] {
   const timing = getParseTiming(options)
+  const tokenizerTiming = timing
+    ? { recordTokenCloneMs: (durationMs: number) => addTiming(timing, 'tokenCloneMs', durationMs) }
+    : undefined
   const parseStartedAt = timing ? getParserNow() : 0
   const isFinal = !!options.final
   // Ensure markdown is a string — guard against null/undefined inputs from callers
   // todo: 下面的特殊 math 其实应该更精确匹配到() 或者 $ $ 或者 \[ \] 内部的内容
   const sourceMarkdown = (markdown ?? '').toString()
-  if (shouldResetTopLevelStreamCacheForFinalAutoParse(md, options)) {
-    md.stream!.reset!()
-    clearTolerantMathBoundaryStreamCache(md)
+  if (resetTopLevelTokenizerForFinalAutoParse(md, options)) {
     // The safe-markdown cache is owned by the top-level streaming session;
     // a final auto-parse ends that session, so drop the retained source +
     // transform (the next stream parse starts a fresh document).
@@ -1637,7 +1086,7 @@ export function parseMarkdownToStructure(
     const preHook = options.preTransformTokens
     const postHook = options.postTransformTokens
     if (shouldUseTopLevelStreamParse(md, options) || typeof preHook === 'function' || typeof postHook === 'function') {
-      const rawTokens = parseTopLevelTokens(md, safeMarkdown, { __markstreamFinal: isFinal }, options) as unknown as MarkdownToken[]
+      const rawTokens = parseTopLevelTokens(md, safeMarkdown, { __markstreamFinal: isFinal }, options, tokenizerTiming) as unknown as MarkdownToken[]
       const hookedTokens = typeof preHook === 'function' ? (preHook(rawTokens) || rawTokens) : rawTokens
       if (typeof postHook === 'function')
         postHook(hookedTokens)
@@ -1647,7 +1096,7 @@ export function parseMarkdownToStructure(
 
   // Get tokens from markdown-it
   const tokenizeStartedAt = timing ? getParserNow() : 0
-  const tokens = parseTopLevelTokens(md, safeMarkdown, { __markstreamFinal: isFinal }, options)
+  const tokens = parseTopLevelTokens(md, safeMarkdown, { __markstreamFinal: isFinal }, options, tokenizerTiming)
   if (timing)
     addTiming(timing, 'tokenizeMs', getParserNow() - tokenizeStartedAt)
   // Defensive: ensure tokens is an array
@@ -1692,7 +1141,10 @@ export function parseMarkdownToStructure(
     __sourceMarkdown: safeMarkdown,
     __customHtmlBlockCursor: 0,
   }
-  let result = processTopLevelTokensWithReuse(md, safeMarkdown, transformedTokens, internalOptions, timing)
+  let result = processTopLevelTokensWithReuse(md, safeMarkdown, transformedTokens, internalOptions, {
+    processTokens: (nextTokens, nextOptions) => processTokensWithTiming(nextTokens, nextOptions, timing),
+    recordReusedTopLevelNodes: count => addTiming(timing, 'processTokensReusedTopLevelNodes', count),
+  })
 
   // Backwards compatible token-level post hook: if provided and returns
   // a modified token array, re-process tokens and override node-level result.
